change after reviweing task for ems 2.6.9.4

+++ ems-dao-api/src/main/java/com/gam/nocr/ems/data/dao/CardDAO.java
@@ -75,5 +75,5 @@ public interface CardDAO extends EmsBaseDAO<CardTO> {
 	 */
 	void unconfirmAllCardsOfBatch(Long batchId) throws BaseException;
 
-    Integer countCardLostDate(Long batchId) throws BaseException;
+    Long countCardLostDate(Long batchId) throws BaseException;
 }
